requirements.txt
boto3
botocore
azure-identity
azure-mgmt-resource
google-auth
google-cloud-compute
python-dotenv